Usando make file e correções a mais

+++ src/exercicio.h
@@ -4,6 +4,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <unistd.h>
+#include <stdint.h>
 
 typedef enum {
     SEQ,
@@ -17,8 +18,8 @@ typedef enum {
 
 void program(void);
 int is_prime(int N);
-int count_primes_seq(int N);
-int count_primes_par(int N, int P, ipc ip);
+uint64_t count_primes_seq(int N);
+uint64_t count_primes_par(int N, int P, ipc ip);
 int* partition_interval(int interval, int partitions);
 
 #endif
